+++ FQHE/src/HilbertSpace/BosonOnSphereShort.h
@@ -240,14 +240,14 @@ class BosonOnSphereShort :  public ParticleOnSphere
   // state = reference to the state to convert
   // reference = set which component has to be normalized to 1
   // return value = converted state
-  virtual RealVector& ConvertToUnnormalizedMonomial(RealVector& state, unsigned int reference = 0);
+  virtual RealVector& ConvertToUnnormalizedMonomial(RealVector& state, long reference = 0);
 
   // convert a state such that its components are now expressed in the normalized basis
   //
   // state = reference to the state to convert
   // reference = set which component has been normalized to 1
   // return value = converted state
-  virtual RealVector& ConvertFromUnnormalizedMonomial(RealVector& state, unsigned int reference = 0);
+  virtual RealVector& ConvertFromUnnormalizedMonomial(RealVector& state, long reference = 0);
 
   // fuse two states which belong to different Hilbert spaces 
   //
